Add a new session
When the user clicks the “new session” button in the navigation bar, a modal will appear. The user can input the name of their new session and then add it to the session log. There is a hidden form field that ensures the default number of pomodoros is 0. If the user decides to cancel making a new session, they can click the cancel button.