Update number of pomodoros
This feature will add one pomodoro every time a user clicks on the button. A PATCH AJAX request runs, and the page is re-rendered to reflect the new number of pomodoros. The user should add a pomodoro after they complete 25 minutes of focused work. When the list of sessions re-renders, it seems to sort by updated date. However, this sort is going in the wrong direction - I need the latest updated session to appear at the top, not the bottom.