Refactored timer
In order to separate the timer from the rest of the page, I placed it in a modal. When the user clicks on the “timer” button in the navigation bar, the timer will appear.
Added a reset button - the reset button will stop the timer, set the duration back to 1500, and set the timer to null.