Fix timer bugs - this time it works, I hope
Trying out another approach using show and hide. Once a user clicks a button, it goes away so they simply don't have the ability to click it again. If they click START, the only options left are to either STOP or RESET. If they click STOP, the only options are to START or RESET. If they click RESET, the only option they have left is to START.